interview-app/backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
google-generativeai>=0.8.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
supabase>=2.0.0,<3.0.0
cloudinary>=1.42.0,<2.0.0
python-multipart>=0.0.9,<1.0.0
